Add openssl to python build properly
Stick with openssl for now as we know that works, not that I don't
think libressl will work as well, but lets not try to change too many
variables at a time, especially not the dependencies.

It looks like we don't need to use --with-openssl to configure of python
when we set the compiler and linker flags properly.