Workaround for sip issues on macos 10.15.4 in qemu
Firstly I got an error with the default sip from somehwere <insert place
here> seemed to be for 32 bit systems, but this qemu image is x86_64.

Then I tried to install sip with pip, but that got me a sip version 5
which is not commandline compatible with the old.

I then discovered that sip from homebrew is still at version 4, so used
this, but I had to force the wxpython build to use that by setting the
environment variable SIP to build.py form wxpython to make it use it.

Now wxpython builds.