Update signing for Apple Silicon
I added arch options to build.py.

On Apple Silicon, we need to sign things before we
run them.

We now sign KiCad with an adhoc signature by default.

Adhoc signatures seem to not have verifiable
timestamps, so we stop checking them in that
case.

If we want to support ad-hoc signing, we have to allow
folks to not enable the Hardened Runtime.