Remove apple.py notarization support
Using altool for notarization meant we had to submit and poll for
completion, while notarytool handles the waiting for us. The shell
equivalent of the apple.py notary subcommand is only a few simple
lines, so I'd rather we didn't maintain it here.

I added information on migrating away in the README.

For https://gitlab.com/kicad/packaging/kicad-mac-builder/-/issues/402