Allow wxpython.org site downtime workarounds
If you set DOXYGEN= and WAF= and point to a Doxygen binary
and a waf file, this should skip downloading Doxygen and WAF through
the wxpython build process.

This includes the waf file that wxpython.org hosted in external/, and
in testing, I used the official Doxygen 1.8.8 DMG and pointed at
/Applications/Doxygen.app/Contents/Resources/doxygen

This also sets WX_SKIP_DOXYGEN_VERSION_CHECK=1.